faceid-master/requirements.txt
PyQt5
pillow
requests
opencv-python
bcolz
matplotlib
numpy
pandas
scikit-learn
tensorboardX
tqdm

fastapi
uvicorn
email_validator
aiofiles
sqlalchemy
mysql-connector-python

pyjwt
passlib[bcrypt]
python-multipart